requirements.txt
requests==2.18.4
esptool==2.1
kivy==1.10.0
pyserial==3.4
